Create wmtk::TriOptimizerMesh by moving triwild's implementation
The 2D half of the shared-base work. Seeded by MOVING TriWildMesh's code rather
than designing an abstraction, so triwild is byte-identical by construction and
any output movement is confined to the later commit where simwild-2D adopts the
base.

Moved: the three attribute types (VertexAttributes, EdgeAttributes,
FaceAttributes), the attribute collections, the envelope/solver/rounding members,
MAX_ENERGY, and 20 leaf methods -- is_inverted x5, get_quality x3, round,
round_all_vertices, get_length2, get_max_avg_energy, partition_mesh{,_morton},
is_edge_on_{surface,bbox} x4, get_edges_by_condition, gradation_smooth_sizing,
vertex_is_on_surface, edge_is_on_surface.

TriWildMesh keeps what is genuinely triwild's: feature points and their
retention checks, force-split and high-valence gating, the winding-number and
flood-fill filters, init_mesh, write_msh_groups, and the operation caches.

The base holds `OptimizerParameters&`; TriWildMesh adds `Parameters& m_tri_params`
for the 12 references to triwild-only fields. VertexAttributes carries
m_feature_id even though simwild-2D has no 0-dimensional features -- an attribute
collection has one element type, and one unused size_t per vertex is cheaper than
splitting the collection.

Verified as a move, not a rewrite: every line removed from TriWildMesh.cpp is
present verbatim in TriOptimizerMesh.cpp modulo the class name, and every removed
header line is accounted for in the new pair.

Gate: ctest 107/107. All 53 registered configs byte-identical (report fields and
the md5 of every emitted mesh), including all 16 triwild ones. tetwild_crown and
tetwild_octocat run 10 threads and are nondeterministic there, so they were
compared separately against a HEAD build at num_threads 0 -- byte-identical.
