+++ components/triwild/wmtk/components/triwild/TriWildMesh.cpp
@@ -38,11 +38,6 @@
 
 namespace wmtk::components::triwild {
 
-VertexAttributes::VertexAttributes(const Vector2r& p)
-{
-    m_pos = p;
-    m_posf = to_double(p);
-}
 
 void TriWildMesh::mesh_improvement(int max_its)
 {
@@ -81,11 +76,11 @@ void TriWildMesh::mesh_improvement(int max_its)
         // whatever the two passes after it turn that state into. Breaking early leaves the
         // remaining passes of this iteration unrun, which is the point: they are not needed.
         double max_energy = 0., avg_energy = 0.;
-        if (!m_params.interleaved_smoothing) {
+        if (!m_tri_params.interleaved_smoothing) {
             std::tie(max_energy, avg_energy) =
-                local_operations({{1, 1, 1, m_params.num_smoothing_passes}});
+                local_operations({{1, 1, 1, m_tri_params.num_smoothing_passes}});
         } else {
-            const int k = m_params.interleaved_smoothing_passes;
+            const int k = m_tri_params.interleaved_smoothing_passes;
             const std::array<std::array<int, 4>, 3> passes = {
                 {{{1, 0, 0, k}}, {{0, 1, 0, k}}, {{0, 0, 1, k}}}};
             for (const std::array<int, 4>& ops : passes) {
@@ -476,7 +471,7 @@ void TriWildMesh::init_mesh(
         m_E_envelope[i] = E_env.row(i);
     }
 
-    m_envelope = std::make_shared<SampleEnvelope>(!m_params.use_sample_envelope);
+    m_envelope = std::make_shared<SampleEnvelope>(!m_tri_params.use_sample_envelope);
     m_envelope->init(m_V_envelope, m_E_envelope, m_envelope_eps);
     logger().info(
         "Envelope: {} (eps {:.6})",
@@ -485,7 +480,7 @@ void TriWildMesh::init_mesh(
 
     // Sanity check: All surface edges must be inside the envelope. Opt-in: see
     // Parameters::check_envelope_at_init for why it is not worth its cost by default.
-    if (m_params.check_envelope_at_init) {
+    if (m_tri_params.check_envelope_at_init) {
         logger().info("Envelope sanity check");
         const auto surf_edges = get_edges_by_condition([](auto& f) { return f.m_is_surface_fs; });
         for (const auto& verts : surf_edges) {
@@ -499,7 +494,7 @@ void TriWildMesh::init_mesh(
     }
 
     // Track the bounding box. This is the box of the *domain* -- the arrangement covers the
-    // input's bounding box grown by the background grid -- which is not m_params.box_min /
+    // input's bounding box grown by the background grid -- which is not m_tri_params.box_min /
     // box_max: those come from the input curves and set the tolerances. Deriving it from V
     // keeps this check self-consistent whatever the input box is.
     const Vector2d domain_min = V.colwise().minCoeff();
@@ -630,14 +625,6 @@ size_t TriWildMesh::refine_sizing_around_worst(double max_energy)
     return refined.size();
 }
 
-void TriWildMesh::gradation_smooth_sizing(double grade, const std::vector<size_t>& seeds)
-{
-    utils::gradation_smooth_sizing(
-        grade,
-        seeds,
-        [this](size_t v) -> double& { return m_vertex_attribute[v].m_sizing_scalar; },
-        [this](size_t v) { return get_one_ring_vids_for_vertex_duplicate(v); });
-}
 
 void TriWildMesh::write_msh_groups(std::string file, const bool write_envelope)
 {
@@ -951,70 +938,6 @@ int TriWildMesh::flood_fill()
     return current_id;
 }
 
-void TriWildMesh::partition_mesh()
-{
-    auto m_vertex_partition_id = partition_TriMesh(*this, NUM_THREADS);
-    for (size_t i = 0; i < m_vertex_partition_id.size(); i++) {
-        m_vertex_attribute[i].partition_id = m_vertex_partition_id[i];
-    }
-}
-
-void TriWildMesh::partition_mesh_morton()
-{
-    if (NUM_THREADS == 0) {
-        return;
-    }
-    logger().info("Number of parts: {} by morton", NUM_THREADS);
-
-    // The shared partitioner is 3D; a zero z leaves the bounding box, the scale and the
-    // Morton code exactly where a 2D-specific version would put them.
-    std::vector<size_t> partition_id;
-    wmtk::partition_vertex_morton(
-        vert_capacity(),
-        [this](size_t i) {
-            const Vector2d& p = m_vertex_attribute[i].m_posf;
-            return Eigen::Vector3d(p[0], p[1], 0);
-        },
-        NUM_THREADS,
-        partition_id);
-
-    for (size_t i = 0; i < partition_id.size(); i++) {
-        m_vertex_attribute[i].partition_id = partition_id[i];
-    }
-}
-
-double TriWildMesh::get_length2(const Tuple& l) const
-{
-    auto& m = *this;
-    auto& v1 = l;
-    auto v2 = l.switch_vertex(m);
-    double length =
-        (m.m_vertex_attribute[v1.vid(m)].m_posf - m.m_vertex_attribute[v2.vid(m)].m_posf)
-            .squaredNorm();
-    return length;
-}
-
-std::tuple<double, double> TriWildMesh::get_max_avg_energy()
-{
-    double max_energy = -1.;
-    double avg_energy = 0.;
-    auto cnt = 0;
-
-    for (int i = 0; i < tri_capacity(); i++) {
-        const Tuple tup = tuple_from_tri(i);
-        if (!tup.is_valid(*this)) {
-            continue;
-        }
-        const double q = m_face_attribute[tup.fid(*this)].m_quality;
-        max_energy = std::max(max_energy, q);
-        avg_energy += q;
-        cnt++;
-    }
-
-    avg_energy /= cnt;
-
-    return std::make_tuple(max_energy, avg_energy);
-}
 
 std::vector<size_t> TriWildMesh::active_vertices() const
 {
@@ -1028,70 +951,6 @@ std::vector<size_t> TriWildMesh::active_vertices() const
         [this](size_t vid) { return m_vertex_attribute[vid].m_is_on_surface; });
 }
 
-bool TriWildMesh::is_inverted_f(const Tuple& loc) const
-{
-    return is_inverted_f(loc.fid(*this));
-}
-
-bool TriWildMesh::is_inverted_f(const size_t fid) const
-{
-    auto vs = oriented_tri_vids(fid);
-
-    igl::predicates::exactinit();
-    auto res = igl::predicates::orient2d(
-        m_vertex_attribute[vs[0]].m_posf,
-        m_vertex_attribute[vs[1]].m_posf,
-        m_vertex_attribute[vs[2]].m_posf);
-    if (res == igl::predicates::Orientation::POSITIVE) {
-        return false;
-    }
-    return true;
-}
-
-bool TriWildMesh::is_inverted(const std::array<size_t, 3>& vs) const
-{
-    // Return a positive value if the point pd lies below the
-    // plane passing through pa, pb, and pc; "below" is defined so
-    // that pa, pb, and pc appear in counterclockwise order when
-    // viewed from above the plane.
-
-    if (m_vertex_attribute[vs[0]].m_is_rounded && m_vertex_attribute[vs[1]].m_is_rounded &&
-        m_vertex_attribute[vs[2]].m_is_rounded) {
-        igl::predicates::exactinit();
-        auto res = igl::predicates::orient2d(
-            m_vertex_attribute[vs[0]].m_posf,
-            m_vertex_attribute[vs[1]].m_posf,
-            m_vertex_attribute[vs[2]].m_posf);
-        if (res == igl::predicates::Orientation::POSITIVE) {
-            return false;
-        }
-        return true;
-    } else {
-        const Vector2r& v0 = m_vertex_attribute[vs[0]].m_pos;
-        const Vector2r& v1 = m_vertex_attribute[vs[1]].m_pos;
-        const Vector2r& v2 = m_vertex_attribute[vs[2]].m_pos;
-        const Vector2r a = v1 - v0;
-        const Vector2r b = v2 - v0;
-        Rational res = a.x() * b.y() - a.y() * b.x();
-        if (res > 0) {
-            return false;
-        } else {
-            return true;
-        }
-    }
-}
-
-bool TriWildMesh::is_inverted(const Tuple& loc) const
-{
-    auto vs = oriented_tri_vids(loc);
-    return is_inverted(vs);
-}
-
-bool TriWildMesh::is_inverted(const size_t fid) const
-{
-    auto vs = oriented_tri_vids(fid);
-    return is_inverted(vs);
-}
 
 std::pair<size_t, size_t> TriWildMesh::feature_retention(double* worst_ratio) const
 {
@@ -1144,7 +1003,7 @@ std::pair<size_t, size_t> TriWildMesh::feature_retention(double* worst_ratio) co
 bool TriWildMesh::smoothing_position_is_allowed(const size_t vid, const Vector2d& p) const
 {
     const size_t f = m_vertex_attribute[vid].m_feature_id;
-    if (f == NO_FEATURE || !m_params.preserve_feature_points) {
+    if (f == NO_FEATURE || !m_tri_params.preserve_feature_points) {
         return true;
     }
     // A ball, not a pin. The vertex may move anywhere within eps of the feature it stands
@@ -1154,7 +1013,7 @@ bool TriWildMesh::smoothing_position_is_allowed(const size_t vid, const Vector2d
 
 bool TriWildMesh::collapse_breaks_feature(const size_t v1_id, const size_t v2_id) const
 {
-    if (!m_params.preserve_feature_points) {
+    if (!m_tri_params.preserve_feature_points) {
         return false;
     }
     // v1 disappears into v2, and v2 does not move. So the question is only whether the
@@ -1177,90 +1036,6 @@ bool TriWildMesh::collapse_breaks_feature(const size_t v1_id, const size_t v2_id
            m_feature_eps * m_feature_eps;
 }
 
-size_t TriWildMesh::round_all_vertices()
-{
-    if (m_all_rounded.load(std::memory_order_relaxed)) {
-        return 0;
-    }
-
-    size_t reclaimed = 0, still_unrounded = 0;
-    for (const Tuple& v : get_vertices()) {
-        if (m_vertex_attribute[v.vid(*this)].m_is_rounded) {
-            continue;
-        }
-        if (round(v)) {
-            ++reclaimed;
-        } else {
-            ++still_unrounded;
-        }
-    }
-
-    if (still_unrounded == 0) {
-        m_all_rounded.store(true, std::memory_order_relaxed);
-    }
-    if (reclaimed > 0 || still_unrounded > 0) {
-        logger().info(
-            "rounding sweep: reclaimed {}, still unrounded {}",
-            reclaimed,
-            still_unrounded);
-    }
-    return reclaimed;
-}
-
-bool TriWildMesh::round(const Tuple& v)
-{
-    size_t i = v.vid(*this);
-    if (m_vertex_attribute[i].m_is_rounded) {
-        return true;
-    }
-
-    auto old_pos = m_vertex_attribute[i].m_pos;
-    m_vertex_attribute[i].m_pos << m_vertex_attribute[i].m_posf[0], m_vertex_attribute[i].m_posf[1];
-    auto conn_tets = get_one_ring_tris_for_vertex(v);
-    m_vertex_attribute[i].m_is_rounded = true;
-    for (const Tuple& tet : conn_tets) {
-        if (is_inverted(tet)) {
-            m_vertex_attribute[i].m_is_rounded = false;
-            m_vertex_attribute[i].m_pos = old_pos;
-            return false;
-        }
-    }
-
-    return true;
-}
-
-double TriWildMesh::get_quality(const std::array<size_t, 3>& vs) const
-{
-    std::array<Vector2d, 3> ps;
-    for (size_t k = 0; k < 3; k++) {
-        ps[k] = m_vertex_attribute[vs[k]].m_posf;
-    }
-    double energy = -1.;
-    {
-        std::array<double, 6> T;
-        for (size_t k = 0; k < 3; k++)
-            for (size_t j = 0; j < 2; j++) {
-                T[k * 2 + j] = ps[k][j];
-            }
-        energy = AMIPS2D_energy(T);
-    }
-    if (std::isinf(energy) || std::isnan(energy) || energy < 2 - 1e-3) {
-        return MAX_ENERGY;
-    }
-    return energy;
-}
-
-double TriWildMesh::get_quality(const Tuple& loc) const
-{
-    auto its = oriented_tri_vids(loc);
-    return get_quality(its);
-}
-
-double TriWildMesh::get_quality(const size_t fid) const
-{
-    auto its = oriented_tri_vids(fid);
-    return get_quality(its);
-}
 
 void TriWildMesh::write_vtu(const std::string& path) const
 {
@@ -1339,60 +1114,5 @@ void TriWildMesh::write_vtu(const std::string& path) const
     }
 }
 
-std::vector<std::array<size_t, 2>> TriWildMesh::get_edges_by_condition(
-    std::function<bool(const EdgeAttributes&)> cond) const
-{
-    std::vector<std::array<size_t, 2>> res;
-    for (const Tuple& e : get_edges()) {
-        size_t eid = e.eid(*this);
-        if (cond(m_edge_attribute[eid])) {
-            res.push_back({{e.vid(*this), e.switch_vertex(*this).vid(*this)}});
-        }
-    }
-    return res;
-}
-
-bool TriWildMesh::is_edge_on_surface(const Tuple& loc) const
-{
-    const auto vs = get_edge_vids(loc);
-    if (!m_vertex_attribute.at(vs[0]).m_is_on_surface ||
-        !m_vertex_attribute.at(vs[1]).m_is_on_surface) {
-        return false;
-    }
-
-    const size_t eid = loc.eid(*this);
-    return m_edge_attribute[eid].m_is_surface_fs;
-}
-bool TriWildMesh::is_edge_on_surface(const std::array<size_t, 2>& vids) const
-{
-    if (!m_vertex_attribute.at(vids[0]).m_is_on_surface ||
-        !m_vertex_attribute.at(vids[1]).m_is_on_surface) {
-        return false;
-    }
-
-    const auto [_, eid] = tuple_from_edge(vids);
-    return m_edge_attribute[eid].m_is_surface_fs;
-}
-bool TriWildMesh::is_edge_on_bbox(const Tuple& loc) const
-{
-    const auto vs = get_edge_vids(loc);
-    if (m_vertex_attribute.at(vs[0]).on_bbox_faces.empty() ||
-        m_vertex_attribute.at(vs[1]).on_bbox_faces.empty()) {
-        return false;
-    }
-
-    const size_t eid = loc.eid(*this);
-    return m_edge_attribute[eid].m_is_bbox_fs >= 0;
-}
-
-bool TriWildMesh::is_edge_on_bbox(const std::array<size_t, 2>& vids) const
-{
-    if (m_vertex_attribute.at(vids[0]).on_bbox_faces.empty() ||
-        m_vertex_attribute.at(vids[1]).on_bbox_faces.empty()) {
-        return false;
-    }
-    const auto [_, eid] = tuple_from_edge(vids);
-    return m_edge_attribute[eid].m_is_bbox_fs >= 0;
-}
 
 } // namespace wmtk::components::triwild
