+++ components/triwild/wmtk/components/triwild/TriWildMesh.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <wmtk/TriOptimizerMesh.h>
+
 #include <wmtk/SurfaceTagAttributes.h>
 #include <cstdlib>
 
@@ -27,75 +29,26 @@
 
 namespace wmtk::components::triwild {
 
+/// The attribute types now live on the shared base; keep the unqualified names working.
+using VertexAttributes = wmtk::TriOptimizerMesh::VertexAttributes;
+using EdgeAttributes = wmtk::TriOptimizerMesh::EdgeAttributes;
+using FaceAttributes = wmtk::TriOptimizerMesh::FaceAttributes;
+
 /// A vertex that stands for no input feature point. See VertexAttributes::m_feature_id.
 inline constexpr size_t NO_FEATURE = std::numeric_limits<size_t>::max();
 
-// TODO: missing comments on what these attributes are
-class VertexAttributes
+class TriWildMesh : public wmtk::TriOptimizerMesh
 {
 public:
-    Vector2d m_posf;
-    Vector2r m_pos;
-    bool m_is_rounded = false;
+    /// The base holds only wmtk::OptimizerParameters; this is the same object, typed, for the
+    /// triwild-only fields (features, high valence, smoothing schedule, box).
+    Parameters& m_tri_params;
 
-    bool m_is_on_surface = false;
-    std::vector<int> on_bbox_faces;
-
-    /**
-     * Index into TriWildMesh::m_feature_points, or NO_FEATURE.
-     *
-     * A feature point is a 0-dimensional feature of the curve network: an open polyline's
-     * endpoint, or a junction. This is the 2D counterpart of tetwild's m_is_on_open_boundary,
-     * with one deliberate difference -- it names WHICH feature the vertex stands for, not
-     * merely that it stands for one.
-     *
-     * That difference is the whole point. tetwild asks "is the survivor inside the envelope
-     * of the boundary?", a containment question, and in 3D a boundary is a curve with many
-     * edges so collapsing one shortens it rather than deleting it. In 2D the feature is a
-     * single point, and a containment test passes trivially when one endpoint is collapsed
-     * onto another -- the target IS a feature point, distance zero -- while the first
-     * endpoint quietly stops being represented. Preserving features is a COVERAGE property,
-     * so the constraint has to bind a vertex to a specific point.
-     */
-    size_t m_feature_id = NO_FEATURE;
-
-    double m_sizing_scalar = 1;
-
-    size_t partition_id = 0;
-
-    VertexAttributes() {}
-    VertexAttributes(const Vector2r& p);
-};
-
-using EdgeAttributes = wmtk::SurfaceTagAttributes;
-
-class FaceAttributes
-{
-public:
-    double m_quality;
-    double m_winding_number = 0;
-    std::set<int64_t> tags;
-    int part_id = -1;
-};
-
-class TriWildMesh : public wmtk::TriMesh
-{
-public:
-    int m_debug_print_counter = 0;
     /// Iterations mesh_improvement actually used. Reported so a run that needs the whole
     /// budget is visible as such, and asserted against in the integration tests.
     int m_iterations_used = 0;
-    size_t m_tags_count = 0;
-    std::map<int64_t, std::string> m_tag_id_to_name;
-    std::map<std::string, int64_t> m_tag_name_to_id;
-
-    const double MAX_ENERGY = 1e50;
-
-    Parameters& m_params;
     std::vector<Vector2d> m_V_envelope;
     std::vector<Vector2i> m_E_envelope;
-    std::shared_ptr<SampleEnvelope> m_envelope;
-    double m_envelope_eps = -1;
     /**
      * Radius of the ball a feature-carrying vertex must stay inside.
      *
@@ -129,9 +82,6 @@ public:
     std::vector<Vector2d> m_feature_points;
     /// Collapses refused because they would drop or displace a feature point. Diagnostic.
     std::atomic<size_t> m_feature_rejects = 0;
-    /// Whether the current collapse pass applies the target-length limit; read by
-    /// collapse_edge_before, which is where that limit is now enforced.
-    bool m_collapse_limit_length = true;
 
     /**
      * @brief May vertex `vid` sit at `p`?
@@ -154,19 +104,6 @@ public:
     /// True iff collapsing v1 into v2 would drop or displace a feature point.
     bool collapse_breaks_feature(const size_t v1_id, const size_t v2_id) const;
 
-    using VertAttCol = wmtk::AttributeCollection<VertexAttributes>;
-    using EdgeAttCol = wmtk::AttributeCollection<EdgeAttributes>;
-    using FaceAttCol = wmtk::AttributeCollection<FaceAttributes>;
-    VertAttCol m_vertex_attribute;
-    EdgeAttCol m_edge_attribute;
-    FaceAttCol m_face_attribute;
-
-    wmtk::threading::enumerable_thread_specific<std::unique_ptr<polysolve::nonlinear::Solver>>
-        m_solver;
-
-    /// Why smoothing attempts were refused, reported once per pass.
-    optimization::SmoothRejectCounters m_smooth_rejects;
-
 
     /// Position hooks for the shared 2D smoothing driver. triwild keeps both a working
     /// double position and an exact rational one, so writing goes through here.
@@ -175,29 +112,18 @@ public:
     std::shared_ptr<SampleEnvelope> smoothing_energy_envelope(const size_t vid) const;
     std::shared_ptr<SampleEnvelope> smoothing_containment_envelope(const size_t vid) const;
 
-    // scaling factors
-    double m_s_amips = -1;
-    double m_s_envelope = -1;
-
     TriWildMesh(Parameters& _m_params, double envelope_eps, int _num_threads = 0)
-        : m_params(_m_params)
-        , m_envelope_eps(envelope_eps)
+        : wmtk::TriOptimizerMesh(_m_params)
+        , m_tri_params(_m_params)
         , m_feature_eps(envelope_eps)
     {
+        m_envelope_eps = envelope_eps;
         NUM_THREADS = _num_threads;
         p_vertex_attrs = &m_vertex_attribute;
         p_edge_attrs = &m_edge_attribute;
         p_face_attrs = &m_face_attribute;
 
         optimization::deactivate_opt_logger();
-
-        m_s_amips = 1.;
-        /**
-         * eps makes it such that the energy is relative to the envelope thickness. As it's a
-         * squared energy, we need eps^2.
-         */
-        m_s_envelope = 1. / (m_params.eps * m_params.eps);
-
 
         double& wa = m_params.w_amips;
         double& we = m_params.w_envelope;
@@ -206,19 +132,6 @@ public:
     }
 
     ~TriWildMesh() {}
-
-    // TODO: this should not be here
-    void partition_mesh();
-
-    // TODO: morton should not be here, but inside wmtk
-    void partition_mesh_morton();
-
-    size_t get_partition_id(const Tuple& loc) const
-    {
-        return m_vertex_attribute[loc.vid(*this)].partition_id;
-    }
-
-    double get_length2(const Tuple& l) const;
 
 public:
     /**
@@ -273,16 +186,6 @@ public:
      */
     size_t refine_sizing_around_worst(double max_energy);
 
-    /**
-     * @brief Monotone (only-decreasing) gradation smoothing of the sizing field.
-     *
-     * Enforces m_sizing_scalar[v] <= grade * m_sizing_scalar[u] for every edge (u,v),
-     * propagating outward from `seeds` with a min-relaxation. It never raises a sizing
-     * value, so it only ever spreads more refinement into the halo around already-refined
-     * vertices, avoiding sharp resolution jumps.
-     */
-    void gradation_smooth_sizing(double grade, const std::vector<size_t>& seeds);
-
     /// The longest edge of each current worst triangle. split_all_edges force-splits
     /// exactly these edges (bypassing the length gate), so a stuck sliver's long edge is
     /// split immediately without changing the sizing field. Populated serially by
@@ -314,9 +217,6 @@ public:
 
     void write_vtu(const std::string& path) const;
 
-    std::vector<std::array<size_t, 2>> get_edges_by_condition(
-        std::function<bool(const EdgeAttributes&)> cond) const;
-
 public:
     void split_all_edges();
     bool split_edge_before(const Tuple& t) override;
@@ -340,61 +240,11 @@ public:
     bool smooth_before(const Tuple& t) override;
     bool smooth_after(const Tuple& t) override;
 
-    /**
-     * @brief A vector containing the vertex position and all positions of the surface neighbors.
-     *
-     * Returns an empty vector if vertex is not on the surface.
-     */
-
-
-    /**
-     * @brief Inversion check using only floating point numbers.
-     */
-    bool is_inverted_f(const Tuple& loc) const;
-    bool is_inverted_f(const size_t fid) const;
-    bool is_inverted(const std::array<size_t, 3>& vs) const;
-    bool is_inverted(const Tuple& loc) const;
-    bool is_inverted(const size_t fid) const;
-    double get_quality(const std::array<size_t, 3>& vs) const;
-    double get_quality(const Tuple& loc) const;
-    double get_quality(const size_t fid) const;
-    bool round(const Tuple& loc);
-
-    /**
-     * @brief Try to round every un-rounded vertex; returns the number reclaimed.
-     *
-     * round() is otherwise only attempted as a side effect of another operation (smoothing
-     * the vertex, or the merged vertex of a collapse), and neither reaches a vertex that
-     * only becomes roundable later: smoothing skips "good" regions by default. Without a
-     * sweep such a vertex keeps exact coordinates into the output for no geometric reason --
-     * and split_edge_after now introduces them unconditionally whenever the rounded midpoint
-     * would invert, so the sweep is what keeps that from reaching the output.
-     *
-     * Skipped outright when m_all_rounded says there is nothing to do.
-     */
-    size_t round_all_vertices();
-
-    /**
-     * @brief True when every vertex is known to be rounded.
-     *
-     * Only trusted when true, and only round_all_vertices() sets it that way. Any code that
-     * leaves a vertex un-rounded must clear it, or the sweep will skip the vertex forever.
-     * Atomic because operations that clear it run in parallel.
-     */
-    std::atomic<bool> m_all_rounded = false;
-    //
-    bool is_edge_on_surface(const Tuple& loc) const;
-    bool is_edge_on_surface(const std::array<size_t, 2>& vids) const;
-    bool is_edge_on_bbox(const Tuple& loc) const;
-    bool is_edge_on_bbox(const std::array<size_t, 2>& vids) const;
-
-    //
     void mesh_improvement(int max_its = 80);
 
     std::tuple<double, double> local_operations(
         const std::array<int, 4>& ops,
         bool collapse_limit_length = true);
-    std::tuple<double, double> get_max_avg_energy();
 
     /**
      * @brief m_quality threshold above which a face is "active" (worth operating on) for
@@ -430,23 +280,6 @@ public:
     void filter_with_flood_fill();
 
     int flood_fill();
-
-    bool vertex_is_on_surface(const size_t vid) const override
-    {
-        return m_vertex_attribute.at(vid).m_is_on_surface ||
-               !m_vertex_attribute.at(vid).on_bbox_faces.empty();
-    }
-    bool edge_is_on_surface(const std::array<size_t, 2>& vids) const override
-    {
-        if (!vertex_is_on_surface(vids[0]) || !vertex_is_on_surface(vids[1])) {
-            return false;
-        }
-
-        const auto [_, eid] = tuple_from_edge(vids);
-        bool on_surface = m_edge_attribute.at(eid).m_is_surface_fs;
-        bool on_bbox = m_edge_attribute.at(eid).m_is_bbox_fs >= 0;
-        return on_surface || on_bbox;
-    }
 
 private:
     ////// Operations
